agy: refactor completer with public actions, add missing subcommands and flags
Split plugin and mcp subcommands into individual files per carapace-bin
pattern. Move complex actions (ActionAgents, ActionModels, ActionPlugins,
ActionMcpServers) to pkg/actions/tools/agy/ so they are exposed as macros
for reuse in specs and other completers.

Added: agy run subcommand, --cwd flag on root, --skip-aliases/--skip-path
on install, --output-format on models/agent, --env/--header/--type flags
on mcp add (using os.ActionEnvironmentVariables and
http.ActionRequestHeaders).
