+++ completers/common/agy_completer/cmd/install.go
@@ -14,5 +14,14 @@ var installCmd = &cobra.Command{
 
 func init() {
 	carapace.Gen(installCmd).Standalone()
+
+	installCmd.Flags().String("dir", "", "Custom install directory")
+	installCmd.Flags().Bool("skip-aliases", false, "Preserve existing agy/antigravity aliases")
+	installCmd.Flags().Bool("skip-path", false, "Do not modify the shell profile PATH")
+
 	rootCmd.AddCommand(installCmd)
+
+	carapace.Gen(installCmd).FlagCompletion(carapace.ActionMap{
+		"dir": carapace.ActionDirectories(),
+	})
 }
