+++ completers/common/agy_completer/cmd/root.go
@@ -2,8 +2,8 @@ package cmd
 
 import (
 	"github.com/carapace-sh/carapace"
+	"github.com/carapace-sh/carapace-bin/pkg/actions/tools/agy"
 	"github.com/spf13/cobra"
-	"strings"
 )
 
 var rootCmd = &cobra.Command{
@@ -27,6 +27,7 @@ func init() {
 	rootCmd.Flags().String("agent", "", "Agent for the current CLI session")
 	rootCmd.Flags().BoolP("continue", "c", false, "Continue the most recent conversation")
 	rootCmd.Flags().String("conversation", "", "Resume a previous conversation by ID")
+	rootCmd.Flags().String("cwd", "", "Working directory for the run")
 	rootCmd.Flags().Bool("dangerously-skip-permissions", false, "Auto-approve all tool permission requests without prompting")
 	rootCmd.Flags().Bool("disable-slash-commands", false, "Disable slash command and skill expansion in print mode")
 	rootCmd.Flags().String("effort", "", "Reasoning effort for the current CLI session (low|medium|high)")
@@ -35,51 +36,27 @@ func init() {
 	rootCmd.Flags().String("log-file", "", "Override CLI log file path")
 	rootCmd.Flags().String("mode", "", "Set the agent execution mode for this session (accept-edits, plan)")
 	rootCmd.Flags().String("model", "", "Model for the current CLI session")
-	rootCmd.Flags().Bool("new-project", false, "Create a new project for this session")
+	rootCmd.Flags().String("new-project", "", "Create a new project with the given name")
 	rootCmd.Flags().String("output-format", "text", "Output format for print mode (text, json, stream-json)")
 	rootCmd.Flags().StringP("print", "p", "", "Run a single prompt non-interactively")
 	rootCmd.Flags().String("print-timeout", "5m0s", "Timeout for print mode wait")
 	rootCmd.Flags().String("project", "", "Project ID or project name for the current CLI session")
 	rootCmd.Flags().String("prompt", "", "Alias for --print")
 	rootCmd.Flags().StringP("prompt-interactive", "i", "", "Run an initial prompt interactively and continue the session")
+	rootCmd.Flags().Bool("remote-control", false, "Start in remote control headless daemon mode")
 	rootCmd.Flags().Bool("sandbox", false, "Run in a sandbox with terminal restrictions enabled")
+	rootCmd.Flags().Bool("version", false, "Show version")
 
 	carapace.Gen(rootCmd).FlagCompletion(carapace.ActionMap{
-		"add-dir": carapace.ActionDirectories(),
-		"agent": carapace.ActionExecCommand("agy", "agents")(func(output []byte) carapace.Action {
-			lines := strings.Split(string(output), "\n")
-			vals := make([]string, 0)
-			for _, line := range lines {
-				if line != "" {
-					parts := strings.Fields(line)
-					if len(parts) > 0 {
-						vals = append(vals, parts[0])
-					}
-				}
-			}
-			return carapace.ActionValues(vals...)
-		}),
-		"effort":       carapace.ActionValues("low", "medium", "high"),
-		"input-format": carapace.ActionValues("text", "stream-json"),
-		"json-schema":  carapace.ActionFiles(".json"),
-		"log-file":     carapace.ActionFiles(),
-		"mode":         carapace.ActionValues("accept-edits", "plan"),
-		"model": carapace.ActionExecCommand("agy", "models")(func(output []byte) carapace.Action {
-			lines := strings.Split(string(output), "\n")
-			vals := make([]string, 0)
-			for _, line := range lines {
-				if line == "" {
-					continue
-				}
-				parts := strings.SplitN(line, "\t", 2)
-				if len(parts) == 2 {
-					vals = append(vals, parts[0], parts[1])
-				} else {
-					vals = append(vals, parts[0], "")
-				}
-			}
-			return carapace.ActionValuesDescribed(vals...)
-		}),
+		"add-dir":       carapace.ActionDirectories(),
+		"agent":         agy.ActionAgents(),
+		"cwd":           carapace.ActionDirectories(),
+		"effort":        carapace.ActionValues("low", "medium", "high"),
+		"input-format":  carapace.ActionValues("text", "stream-json"),
+		"json-schema":   carapace.ActionFiles(".json"),
+		"log-file":      carapace.ActionFiles(),
+		"mode":          carapace.ActionValues("accept-edits", "plan"),
+		"model":         agy.ActionModels(),
 		"output-format": carapace.ActionValues("text", "json", "stream-json"),
 		"print-timeout": carapace.ActionValues("30s", "1m", "10m", "1h").Usage("duration (e.g., 5m, 1h30m)"),
 	})
