+++ completers/common/agy_completer/cmd/run.go
@@ -1,0 +1,24 @@
+package cmd
+
+import (
+	"github.com/carapace-sh/carapace"
+	"github.com/spf13/cobra"
+)
+
+var runCmd = &cobra.Command{
+	Use:     "run",
+	GroupID: "engine",
+	Short:   "Run the agent via the run subcommand",
+	Run:     func(cmd *cobra.Command, args []string) {},
+}
+
+func init() {
+	carapace.Gen(runCmd).Standalone()
+
+	runCmd.Flags().String("cwd", "", "Working directory for the run")
+	rootCmd.AddCommand(runCmd)
+
+	carapace.Gen(runCmd).FlagCompletion(carapace.ActionMap{
+		"cwd": carapace.ActionDirectories(),
+	})
+}
